slim-msm-ngd: Disable core irq in SSR/PDR Down notification
Slimbus core interrupt is getting fired during a ADSP power up(Sub-sys up)
orafter a adsp shutdown(Sub-sys Down). At this point adsp slimbus hardware
is not ON(clocks are off) which is leading to a unclocked access when
HLOS slimbus tried to read the interrupt status register in isr.

Fix is to disable core irq as a part of SSR down notification and enable
it as a part of SSR UP notification in ngd_slim_enable API. And also add a
defensive check in ISR to not access registers if slimbus is
runtime suspended.

Change-Id: I84a599551b98c8580bdfc889ebbe99f48c649895
Signed-off-by: Chandana Kishori Chiluveru <cchiluve@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>