mm/mremap: hold the rmap lock in write mode when moving page table entries
To avoid a race between rmap walk and mremap, mremap does
take_rmap_locks().  The lock was taken to ensure that rmap walk don't miss
a page table entry due to PTE moves via move_pagetables().  The kernel
does further optimization of this lock such that if we are going to find
the newly added vma after the old vma, the rmap lock is not taken.  This
is because rmap walk would find the vmas in the same order and if we don't
find the page table attached to older vma we would find it with the new
vma which we would iterate later.

Change-Id: I11af623e4834e57d0c00de597512b5c787ef0979
Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>
Git-commit: 97113eb39fa7972722ff490b947d8af023e1f6a2
Git-repo: https://android.googlesource.com/kernel/common/
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Ashok Gummadidala <agumma@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>